removed confusing outlet stuff
