+++ frontend/src/components/HomeFeed/HomeFeed.js
@@ -4,10 +4,18 @@ import Post from "../Post/Post";
 import styles from "./HomeFeed.module.css";
 import { posts } from "../../test-data/post-data.js";
 import { user } from "../../test-data/user-data.js";
-import { useOutletContext } from "react-router-dom";
+import ImageModal from "../ImageModal/ImageModal";
 
 const HomeFeed = () => {
-  const [handleImageClick] = useOutletContext();
+  const [isModalOpen, setModalOpen] = useState(false);
+  const [modalImages, setModalImages] = useState([]);
+  const [modalIndex, setModalIndex] = useState(0);
+
+  const handleImageClick = (images, index) => {
+    setModalImages(images);
+    setModalIndex(index);
+    setModalOpen(true);
+  };
 
   return (
     <>
@@ -15,6 +23,14 @@ const HomeFeed = () => {
       {posts.map((post, index) => (
         <Post key={index} post={post} handlePostImageClick={handleImageClick} />
       ))}
+      {isModalOpen && (
+        <ImageModal
+          onOverlayClick={() => setModalOpen(false)}
+          onClose={() => setModalOpen(false)}
+          images={modalImages}
+          index={modalIndex}
+        />
+      )}
     </>
   );
 };
