+++ frontend/src/components/Profile/Profile.js
@@ -1,14 +1,22 @@
-import React from "react";
+import React, { useState } from "react";
 import styles from "./Profile.module.css";
 import ProfileHeader from "./ProfileHeader/ProfileHeader";
 import Post from "../Post/Post";
 import { user } from "../../test-data/user-data";
 import { posts } from "../../test-data/post-data";
 import { AiOutlinePlus } from "react-icons/ai";
-import { useOutletContext } from "react-router-dom";
+import ImageModal from "../ImageModal/ImageModal";
 
 const Profile = () => {
-  const [handleImageClick] = useOutletContext();
+  const [isModalOpen, setModalOpen] = useState(false);
+  const [modalImages, setModalImages] = useState([]);
+  const [modalIndex, setModalIndex] = useState(0);
+
+  const handleImageClick = (images, index) => {
+    setModalImages(images);
+    setModalIndex(index);
+    setModalOpen(true);
+  };
 
   // Fetch data using route path and params
   return (
@@ -26,6 +34,14 @@ const Profile = () => {
       <button className={styles["create-post"]}>
         <AiOutlinePlus />
       </button>
+      {isModalOpen && (
+        <ImageModal
+          onOverlayClick={() => setModalOpen(false)}
+          onClose={() => setModalOpen(false)}
+          images={modalImages}
+          index={modalIndex}
+        />
+      )}
     </div>
   );
 };
